// src/orvalApi/model/authResponseDto.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * backend
 * OpenAPI spec version: 1.0.0
 */
import type { DefaultSelectionPrisma36UserPayload } from "./defaultSelectionPrisma36UserPayload";
import type { AuthTokensDto } from "./authTokensDto";

export interface AuthResponseDto {
  /** User information */
  user: DefaultSelectionPrisma36UserPayload;
  /** Authentication tokens */
  tokens: AuthTokensDto;
  [key: string]: unknown;
}